Rework exports of the `hooks-*` modules.
